LayoutTests/fast/forms/autocomplete-tokens-expected.txt

Tests for valid autocomplete tokens on input, select, and textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Valid keywords:
PASS input.autocomplete is "name"
PASS input.autocomplete is "honorific-prefix"
PASS input.autocomplete is "given-name"
PASS input.autocomplete is "additional-name"
PASS input.autocomplete is "family-name"
PASS input.autocomplete is "honorific-suffix"
PASS input.autocomplete is "nickname"
PASS input.autocomplete is "organization-title"
PASS input.autocomplete is "username"
PASS input.autocomplete is "new-password"
PASS input.autocomplete is "current-password"
PASS input.autocomplete is "organization"
PASS input.autocomplete is "street-address"
PASS input.autocomplete is "address-line1"
PASS input.autocomplete is "address-line2"
PASS input.autocomplete is "address-line3"
PASS input.autocomplete is "address-level4"
PASS input.autocomplete is "address-level3"
PASS input.autocomplete is "address-level2"
PASS input.autocomplete is "address-level1"
PASS input.autocomplete is "country"
PASS input.autocomplete is "country-name"
PASS input.autocomplete is "postal-code"
PASS input.autocomplete is "cc-name"
PASS input.autocomplete is "cc-given-name"
PASS input.autocomplete is "cc-additional-name"
PASS input.autocomplete is "cc-family-name"
PASS input.autocomplete is "cc-number"
PASS input.autocomplete is "cc-exp"
PASS input.autocomplete is "cc-exp-month"
PASS input.autocomplete is "cc-exp-year"
PASS input.autocomplete is "cc-csc"
PASS input.autocomplete is "cc-type"
PASS input.autocomplete is "transaction-currency"
PASS input.autocomplete is "transaction-amount"
PASS input.autocomplete is "language"
PASS input.autocomplete is "bday"
PASS input.autocomplete is "bday-day"
PASS input.autocomplete is "bday-month"
PASS input.autocomplete is "bday-year"
PASS input.autocomplete is "sex"
PASS input.autocomplete is "url"
PASS input.autocomplete is "photo"
PASS input.autocomplete is "tel"
PASS input.autocomplete is "tel-country-code"
PASS input.autocomplete is "tel-national"
PASS input.autocomplete is "tel-area-code"
PASS input.autocomplete is "tel-local"
PASS input.autocomplete is "tel-local-prefix"
PASS input.autocomplete is "tel-local-suffix"
PASS input.autocomplete is "tel-extension"
PASS input.autocomplete is "email"
PASS input.autocomplete is "impp"

Contact scope:
PASS input.autocomplete is "home tel"
PASS input.autocomplete is "home tel-country-code"
PASS input.autocomplete is "home tel-national"
PASS input.autocomplete is "home tel-area-code"
PASS input.autocomplete is "home tel-local"
PASS input.autocomplete is "home tel-local-prefix"
PASS input.autocomplete is "home tel-local-suffix"
PASS input.autocomplete is "home tel-extension"
PASS input.autocomplete is "home email"
PASS input.autocomplete is "home impp"
PASS input.autocomplete is "work tel"
PASS input.autocomplete is "work tel-country-code"
PASS input.autocomplete is "work tel-national"
PASS input.autocomplete is "work tel-area-code"
PASS input.autocomplete is "work tel-local"
PASS input.autocomplete is "work tel-local-prefix"
PASS input.autocomplete is "work tel-local-suffix"
PASS input.autocomplete is "work tel-extension"
PASS input.autocomplete is "work email"
PASS input.autocomplete is "work impp"
PASS input.autocomplete is "mobile tel"
PASS input.autocomplete is "mobile tel-country-code"
PASS input.autocomplete is "mobile tel-national"
PASS input.autocomplete is "mobile tel-area-code"
PASS input.autocomplete is "mobile tel-local"
PASS input.autocomplete is "mobile tel-local-prefix"
PASS input.autocomplete is "mobile tel-local-suffix"
PASS input.autocomplete is "mobile tel-extension"
PASS input.autocomplete is "mobile email"
PASS input.autocomplete is "mobile impp"
PASS input.autocomplete is "fax tel"
PASS input.autocomplete is "fax tel-country-code"
PASS input.autocomplete is "fax tel-national"
PASS input.autocomplete is "fax tel-area-code"
PASS input.autocomplete is "fax tel-local"
PASS input.autocomplete is "fax tel-local-prefix"
PASS input.autocomplete is "fax tel-local-suffix"
PASS input.autocomplete is "fax tel-extension"
PASS input.autocomplete is "fax email"
PASS input.autocomplete is "fax impp"
PASS input.autocomplete is "pager tel"
PASS input.autocomplete is "pager tel-country-code"
PASS input.autocomplete is "pager tel-national"
PASS input.autocomplete is "pager tel-area-code"
PASS input.autocomplete is "pager tel-local"
PASS input.autocomplete is "pager tel-local-prefix"
PASS input.autocomplete is "pager tel-local-suffix"
PASS input.autocomplete is "pager tel-extension"
PASS input.autocomplete is "pager email"
PASS input.autocomplete is "pager impp"

Mode:
PASS input.autocomplete is "shipping street-address"
PASS input.autocomplete is "shipping address-line1"
PASS input.autocomplete is "shipping address-line2"
PASS input.autocomplete is "shipping address-line3"
PASS input.autocomplete is "shipping address-level4"
PASS input.autocomplete is "shipping address-level3"
PASS input.autocomplete is "shipping address-level2"
PASS input.autocomplete is "shipping address-level1"
PASS input.autocomplete is "shipping country"
PASS input.autocomplete is "shipping country-name"
PASS input.autocomplete is "shipping postal-code"
PASS input.autocomplete is "billing street-address"
PASS input.autocomplete is "billing address-line1"
PASS input.autocomplete is "billing address-line2"
PASS input.autocomplete is "billing address-line3"
PASS input.autocomplete is "billing address-level4"
PASS input.autocomplete is "billing address-level3"
PASS input.autocomplete is "billing address-level2"
PASS input.autocomplete is "billing address-level1"
PASS input.autocomplete is "billing country"
PASS input.autocomplete is "billing country-name"
PASS input.autocomplete is "billing postal-code"

Section:
PASS input.autocomplete is "section-parent name"
PASS input.autocomplete is "section-parent honorific-prefix"
PASS input.autocomplete is "section-parent given-name"
PASS input.autocomplete is "section-parent additional-name"
PASS input.autocomplete is "section-parent family-name"
PASS input.autocomplete is "section-parent honorific-suffix"
PASS input.autocomplete is "section-parent nickname"
PASS input.autocomplete is "section-parent organization-title"
PASS input.autocomplete is "section-parent username"
PASS input.autocomplete is "section-parent new-password"
PASS input.autocomplete is "section-parent current-password"
PASS input.autocomplete is "section-parent organization"
PASS input.autocomplete is "section-parent street-address"
PASS input.autocomplete is "section-parent address-line1"
PASS input.autocomplete is "section-parent address-line2"
PASS input.autocomplete is "section-parent address-line3"
PASS input.autocomplete is "section-parent address-level4"
PASS input.autocomplete is "section-parent address-level3"
PASS input.autocomplete is "section-parent address-level2"
PASS input.autocomplete is "section-parent address-level1"
PASS input.autocomplete is "section-parent country"
PASS input.autocomplete is "section-parent country-name"
PASS input.autocomplete is "section-parent postal-code"
PASS input.autocomplete is "section-parent cc-name"
PASS input.autocomplete is "section-parent cc-given-name"
PASS input.autocomplete is "section-parent cc-additional-name"
PASS input.autocomplete is "section-parent cc-family-name"
PASS input.autocomplete is "section-parent cc-number"
PASS input.autocomplete is "section-parent cc-exp"
PASS input.autocomplete is "section-parent cc-exp-month"
PASS input.autocomplete is "section-parent cc-exp-year"
PASS input.autocomplete is "section-parent cc-csc"
PASS input.autocomplete is "section-parent cc-type"
PASS input.autocomplete is "section-parent transaction-currency"
PASS input.autocomplete is "section-parent transaction-amount"
PASS input.autocomplete is "section-parent language"
PASS input.autocomplete is "section-parent bday"
PASS input.autocomplete is "section-parent bday-day"
PASS input.autocomplete is "section-parent bday-month"
PASS input.autocomplete is "section-parent bday-year"
PASS input.autocomplete is "section-parent sex"
PASS input.autocomplete is "section-parent url"
PASS input.autocomplete is "section-parent photo"
PASS input.autocomplete is "section-parent tel"
PASS input.autocomplete is "section-parent tel-country-code"
PASS input.autocomplete is "section-parent tel-national"
PASS input.autocomplete is "section-parent tel-area-code"
PASS input.autocomplete is "section-parent tel-local"
PASS input.autocomplete is "section-parent tel-local-prefix"
PASS input.autocomplete is "section-parent tel-local-suffix"
PASS input.autocomplete is "section-parent tel-extension"
PASS input.autocomplete is "section-parent email"
PASS input.autocomplete is "section-parent impp"

Invalid combinations:
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""
PASS input.autocomplete is ""

Non-autocompleteable controls:
PASS button.autocomplete is "invalid"
PASS keygen.autocomplete is "invalid"
PASS successfullyParsed is true

TEST COMPLETE
